convert osf/metrics.py into a package
