+++ osf/metrics/__init__.py
@@ -1,0 +1,20 @@
+from .preprint_metrics import (
+    PreprintView,
+    PreprintDownload,
+)
+
+from .institution_metrics import (
+    InstitutionProjectCounts,
+    UserInstitutionProjectCounts,
+)
+
+from .registry_metrics import RegistriesModerationMetrics
+
+
+__all__ = (
+    'InstitutionProjectCounts',
+    'PreprintView',
+    'PreprintDownload',
+    'RegistriesModerationMetrics',
+    'UserInstitutionProjectCounts',
+)
